Fix host resolution problem
There was a small bug where hostnames were not being detected when we used domains with more than 3 parts, like Azure's, the splitting was done but
the handler would not identify the host properly.

Using HAR the domain is <name>.<uuid>.<location>.aksapp.io which is 5 parts when splitted. I changed the rule so the host is not discarded.